2021.04.16 14:30 검색, 게시글 수정 사진넣어서 구현, 주석 추가

+++ src/main/java/com/hanghae/market/service/BoardService.java
@@ -30,31 +30,60 @@ public class BoardService {
         return mainDtoList;
     }
 
+    public List<BoardMainDto> getSearchBoard(String title) {
+        List<Board> board = boardRepository.findByTitleContainingOrContentContaining(title, title);
+        List<BoardMainDto> mainDtoList = new ArrayList<>();
+        // main에 필요한 값들만 Dto로 만들어서 보내준다.
+        for(int i=0; i<board.size(); i++){
+            BoardMainDto mainDto = new BoardMainDto(board.get(i));
+            mainDtoList.add(mainDto);
+        }
+        return mainDtoList;
+    }
+
     public void createBoard(BoardRequestDto requestDto, Long userId) {
         User user = userRepository.findById(userId).orElseThrow(
                 () -> new IllegalArgumentException("계정이 존재하지 않습니다.")
         );
         Board board = new Board(requestDto);
-        board.addUser(user);
+        board.addUser(user); // 연관관계 추가 (댓글 만들어지면 또 댓글도 추가해야함)
         boardRepository.save(board);
 
     }
 
     @Transactional
-    public Board updateBoard(Long boardId, BoardRequestDto requestDto) {
+    public Board updateBoard(Long boardId, BoardRequestDto requestDto, Long userId) {
+        User user = userRepository.findById(userId).orElseThrow(
+                () -> new IllegalArgumentException("계정이 존재하지 않습니다.")
+        );
         Board board = boardRepository.findById(boardId).orElseThrow(
                 ()-> new IllegalArgumentException("게시글이 존재하지 않습니다.")
         );
-        board.update(requestDto);
-        return board;
+        if (board.getUser().getId().equals(userId)){
+            board.update(requestDto);
+            return board;
+        }
+        else{
+            return null;
+        }
+
     }
 
     @Transactional
-    public void deleteBoard(Long boardId) {
+    public Board deleteBoard(Long boardId, Long userId) {
+        User user = userRepository.findById(userId).orElseThrow(
+                () -> new IllegalArgumentException("계정이 존재하지 않습니다.")
+        );
         Board board = boardRepository.findById(boardId).orElseThrow(
                 ()-> new IllegalArgumentException("게시글이 존재하지 않습니다.")
         );
-        boardRepository.deleteById(boardId);
+        if (board.getUser().getId().equals(userId)) {
+            boardRepository.deleteById(boardId);
+            return board;
+        }
+        else{
+            return null;
+        }
     }
 
 
@@ -64,4 +93,6 @@ public class BoardService {
         );
         return board;
     }
+
+
 }
